Fix all x86-64 warnings in the main DOSEMU code. 64bit DOSEMU links now.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1598 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/arch/linux/debugger/dosdebug.c
@@ -38,7 +38,7 @@
 
 #define MHP_BUFFERSIZE 8192
 
-#define FOREVER ((((unsigned long)-1L) >> 1) / CLOCKS_PER_SEC)
+#define FOREVER ((((unsigned int)-1) >> 1) / CLOCKS_PER_SEC)
 #define KILL_TIMEOUT 2
 
 fd_set readfds;
